backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
alembic
psycopg2-binary
python-dotenv
pytest
httpx
pydantic[email]
python-jose[cryptography]
passlib==1.7.4
bcrypt<4
